Add constant pool.
