+++ README.md
@@ -52,6 +52,7 @@ Then run the following command to detect the memory leak of the binary version p
 
 - [x] Basic Wasm interpreter on fibonacci;
 - [x] Cache based opcode&immediates optimization;
+- [x] Constant pool optimization;
 - [ ] JIT compiler based optimization;
 - [ ] Stage based full-path type validation;
 - [ ] Full Wasm MVP proposals support;
